Refactor: Implement repository module and apply to views
This commit refactors the application by introducing a `RepositoryModule` for dependency injection of the `NutriFindRepository`. The `DefaultNutriFindRepository` is now provided as the implementation of the `NutriFindRepository` interface.

Additionally, the commit modifies several UI components (`HomeScreen.kt`, `FoodDetailsScreen.kt`, `CategoryViewModel.kt`, `FoodDetailsViewModel.kt`, `HomeViewModel.kt`) to use the `NutriFindRepository` interface instead of the concrete `DefaultNutriFindRepository` implementation. This promotes better abstraction and testability.

The commit also updates the `HomeUiState` and `FoodDetailsUiState` to use `DataResponse` for handling data states, improving the representation of asynchronous data fetching.

Finally, a small change is made in `DataStoreModule.kt` to make the `dataStore` variable public for potential external use.